+++ tests/teststops.cpp
@@ -63,7 +63,7 @@ int pulse( void *outputBuffer, void * /*inputBuffer*/, unsigned int nBufferFrame
   if ( status ) std::cout << "Stream over/underflow detected!" << std::endl;
 
   for ( i=0; i<nBufferFrames; i++ ) {
-    if ( data->frameCounter % data->pulseCount == 0 ) sample = 0.9;
+    if ( data->frameCounter % data->pulseCount == 0 ) sample = 0.9f;
     else sample = 0.0;
     for ( j=0; j<data->channels; j++ )
       *buffer++ = sample;
@@ -107,8 +107,8 @@ int main( int argc, char *argv[] )
   // Let RtAudio print messages to stderr.
   adc->showWarnings( true );
 
-  runtime = RUNTIME * 1000;
-  pausetime = PAUSETIME * 1000;
+  runtime = static_cast<unsigned int>(RUNTIME * 1000);
+  pausetime = static_cast<unsigned int>(PAUSETIME * 1000);
 
   // Set our stream parameters for a duplex stream.
   bufferFrames = 512;
@@ -127,7 +127,7 @@ int main( int argc, char *argv[] )
     oParams.deviceId = adc->getDefaultOutputDevice();
 
   // First, test external stopStream() calls.
-  mydata.pulseCount = PULSE_RATE * fs;
+  mydata.pulseCount = static_cast<unsigned int>(PULSE_RATE * fs);
   mydata.nFrames = 50 * fs;
   mydata.returnValue = 0;
   try {
